+++ internal/ui/search.go
@@ -101,11 +101,6 @@ func (m *Model) handleIMAPSearchResult(msg imapSearchResultMsg) (tea.Model, tea.
 	}
 	m.imapSearchResults = true
 	m.offTabFolder = "Search"
-	// Prepend folder name to subject so user can see where each result is from
-	for i := range msg.emails {
-		folder := msg.emails[i].Folder
-		msg.emails[i].Subject = "[" + folder + "] " + msg.emails[i].Subject
-	}
 	m.emails = msg.emails
 	m.markedUIDs = make(map[uint32]bool)
 	m.filterActive = false
@@ -151,10 +146,6 @@ func (m *Model) handleEverythingResult(msg everythingResultMsg) (tea.Model, tea.
 		return m, nil
 	}
 	m.offTabFolder = "Everything"
-	// Prepend folder name so user knows where each email is
-	for i := range msg.emails {
-		msg.emails[i].Subject = "[" + msg.emails[i].Folder + "] " + msg.emails[i].Subject
-	}
 	m.emails = msg.emails
 	m.markedUIDs = make(map[uint32]bool)
 	m.filterActive = false
@@ -224,9 +215,6 @@ func (m *Model) handleConversationResult(msg conversationResultMsg) (tea.Model, 
 		return m, nil
 	}
 	m.offTabFolder = "Thread"
-	for i := range msg.emails {
-		msg.emails[i].Subject = "[" + msg.emails[i].Folder + "] " + msg.emails[i].Subject
-	}
 	m.emails = msg.emails
 	m.markedUIDs = make(map[uint32]bool)
 	m.filterActive = false